feat(donusum): adim 4 - sp_Prog_BelgeDonusum_Kaydet
Planin BaslikOlustur + SatirAktar adimlarinin birlesimi. Belge uretimi icin
YENI NESNE YOK: gecici tablolardan Kaydet JSON'unu kurup kanonik
sp_Api_Belge_Kaydet_Json'u cagiriyor.

ROTA POLITIKASI fn_Prog_BelgeDonusum_Rota'dan okunuyor: depo kaynagi (YOK ->
NULL, VARSAYILAN7 -> konsinye deposu), doviz (RAPORDOVIZ kaynaktan aynen,
FATURADOVIZI rota politikasina gore), StokDurumDegis, EkipmanSabit,
KdvMuafiyetKopyala, Carpan.

BELGE NUMARASI ARTIK SUNUCUDA: OTOMATIK rotalarda numara TRANSACTION ICINDE
sp_BelgeNoGetir ile tahsis ediliyor. Uygulamada uretilip gonderilmiyor - es
zamanli iki donusumde numara cakismasi riski bitti (degerlendirme raporu
madde 4).

Mevcut belgeye eklemede baslikta YALNIZ ID gonderiliyor; hedef basligin hicbir
alani degismiyor. SatirModu 'delta' (gonderilmeyen satirlara dokunulmaz).

Satir eslesmesi YERI/YERID bagindan okunuyor - Kaydet'in sonuc JSON'u da
Sira->ID veriyor ama INSERT...EXEC ic ice gecmedigi icin yakalanamiyor.

428 KAPSAM DISI (Destek=0): Kaydet yalnizca FATBASLIK/FATURA yaziyor; hedefi
SIPARIS olan bu rota icin sipariş tarafinda es deger kanonik bir kaydet SP'si
yok. Ikiz uretim yazmamak icin rota eski akista birakildi; sipariş kaydet SP'si
yazilinca Destek=1 olacak.

BUG FIX (adim 1'e): Kaydet'te Mf alani INT bildirilmisti, FATURA.MF ise
decimal(24,6) - '0.000000' degeri INT'e cevrilemiyordu. DECIMAL(18,6) yapildi.

TEST (BILIM, ROLLBACK'li) - Dogrula + Kaydet zinciri:
 - hedef belge 114080 olustu, numara transaction icinde tahsis edildi (104151)
 - baslik: TUR 14, KOCANNO 1401, SENARYO 1, EFATURADURUM/EFATURASONUC 0,
   BASLIK/VD/SATICIKODU kaynaktan kopyalandi
 - satirlar: YERI=409, YERID dogru, adet/miktar/tutar dogru, STOKDURUMDEGIS=1
 - #DonusumSatirEsleme doldu
 - kaynak satirlarin KALANI 0'a dusru - donusum bagi calisiyor, ayni satir
   ikinci kez donusturulemiyor

+++ Entegra/GenUpdate/GenDepoUpdate88.sql
@@ -19,7 +19,8 @@
 --   Yeri, YerId  <-- DONUSUM BAGI. Hedef satirin hangi donusum turuyle hangi
 --                    kaynak satirdan uretildigini tutar; kalan hesabinin
 --                    (KalanGrubu) dayandigi tek bilgi budur.
---   PozNo, StokDurumDegis, EkipmanId, Mf, MuhKodu, Kasa, OtvYuzde, OtvMiktar,
+--   PozNo, StokDurumDegis, EkipmanId, Mf (decimal - FATURA.MF ondalikli), MuhKodu,
+--   Kasa, OtvYuzde, OtvMiktar,
 --   IzlemeKodu, Vade, KampanyaId, KdvMuafiyeti
 --
 --   KAPSAM DISI - EN/BOY/YUZEY/SAYI ve TESLIMTARIHI:
@@ -134,7 +135,7 @@ BEGIN
         --   Yeri/YerId DONUSUM BAGIDIR: hedef satir hangi donusum turuyle
         --   hangi kaynak satirdan uretildi. Kalan hesabi buna dayanir.
         Yeri INT NULL, YerId INT NULL, PozNo INT NULL, StokDurumDegis INT NULL,
-        EkipmanId INT NULL, Mf INT NULL, MuhKodu NVARCHAR(50) NULL, Kasa INT NULL,
+        EkipmanId INT NULL, Mf DECIMAL(18,6) NULL, MuhKodu NVARCHAR(50) NULL, Kasa INT NULL,
         OtvYuzde FLOAT NULL, OtvMiktar DECIMAL(18,6) NULL, IzlemeKodu NVARCHAR(50) NULL,
         Vade INT NULL, KampanyaId INT NULL, KdvMuafiyeti INT NULL);
     INSERT @S (Sira, SatirId, Sil, UrunId, Tur, Adet, Miktar, Birim, BirimFiyat, Tutar, Kdv,
@@ -161,7 +162,8 @@ BEGIN
                DovizBirimFiyat DECIMAL(18,6), DovizKurDegeri MONEY, DovizTutari DECIMAL(18,6),
                Aciklama NVARCHAR(250), ProjeId INT, MasrafId INT, OzelKod NVARCHAR(50),
                OzelKod2 NVARCHAR(50), Izleme INT,
-                   Yeri INT, YerId INT, PozNo INT, StokDurumDegis INT, EkipmanId INT, Mf INT,
+                   Yeri INT, YerId INT, PozNo INT, StokDurumDegis INT, EkipmanId INT,
+                   Mf DECIMAL(18,6),
                    MuhKodu NVARCHAR(50), Kasa INT, OtvYuzde FLOAT, OtvMiktar DECIMAL(18,6),
                    IzlemeKodu NVARCHAR(50), Vade INT, KampanyaId INT, KdvMuafiyeti INT,
                    SeriLot NVARCHAR(MAX) AS JSON) J;
